app7DataMasking/microservice2/requirements.txt
uvicorn
fastapi
pydantic
aiohttp
asyncio
aiosqlite
pycryptodome